// types/cake-log/log-detail.ts
export interface LogData {
  cakeLogid: number;
  username: string;
  userProfile: string;
  createAt: string;
  cakeShopId: number;
  cakeShopName: string;
  cakeCategoryName: string;
  title: string;
  thumbnailImage: string;
  body: string;
  isPublic: boolean;
  imageList: string[];
  isScraped: boolean;
}